client/src/ changes: moved Stripe payment button from App.js to Footer.js, added css styling to make it look similar to existing button
